+++ internal/logger/logger.go
@@ -1,4 +1,4 @@
-package iavl
+package logger
 
 import (
 	"fmt"
@@ -8,7 +8,7 @@ var (
 	debugging = false
 )
 
-func debug(format string, args ...interface{}) {
+func Debug(format string, args ...interface{}) {
 	if debugging {
 		fmt.Printf(format, args...)
 	}
