+++ mutable_tree.go
@@ -10,6 +10,8 @@ import (
 	"github.com/pkg/errors"
 
 	dbm "github.com/tendermint/tm-db"
+
+	"github.com/cosmos/iavl/internal/logger"
 )
 
 // ErrVersionDoesNotExist is returned if a requested version does not exist.
@@ -529,13 +531,13 @@ func (tree *MutableTree) SaveVersion() ([]byte, int64, error) {
 	if tree.root == nil {
 		// There can still be orphans, for example if the root is the node being
 		// removed.
-		debug("SAVE EMPTY TREE %v\n", version)
+		logger.Debug("SAVE EMPTY TREE %v\n", version)
 		tree.ndb.SaveOrphans(version, tree.orphans)
 		if err := tree.ndb.SaveEmptyRoot(version); err != nil {
 			return nil, 0, err
 		}
 	} else {
-		debug("SAVE TREE %v\n", version)
+		logger.Debug("SAVE TREE %v\n", version)
 		tree.ndb.SaveBranch(tree.root)
 		tree.ndb.SaveOrphans(version, tree.orphans)
 		if err := tree.ndb.SaveRoot(tree.root, version); err != nil {
@@ -587,7 +589,7 @@ func (tree *MutableTree) SetInitialVersion(version uint64) {
 // DeleteVersions deletes a series of versions from the MutableTree.
 // Deprecated: please use DeleteVersionsRange instead.
 func (tree *MutableTree) DeleteVersions(versions ...int64) error {
-	debug("DELETING VERSIONS: %v\n", versions)
+	logger.Debug("DELETING VERSIONS: %v\n", versions)
 
 	if len(versions) == 0 {
 		return nil
@@ -640,7 +642,7 @@ func (tree *MutableTree) DeleteVersionsRange(fromVersion, toVersion int64) error
 // DeleteVersion deletes a tree version from disk. The version can then no
 // longer be accessed.
 func (tree *MutableTree) DeleteVersion(version int64) error {
-	debug("DELETE VERSION: %d\n", version)
+	logger.Debug("DELETE VERSION: %d\n", version)
 
 	if err := tree.deleteVersion(version); err != nil {
 		return err
